- deleted a lot of validation on the person tryparse function -  changed the call to the tryparse to the right call - deleted  the tryparse in ContactsCenter and placed it into the person Class
